Update celowork.sol
Nice Dapp ayloMac and idea for the CeloWork. However as Ichigo pointed out, the link on your submission is to the live demo and not to the "GitHub Repo" (GH repo). This is the link to the GH repo https://github.com/taylormcneal/CeloWork. So i deleted the testing.sol file from the contract dir as it is not needed, next i worked on the frontend by moving all constants variables to the utils dir. I removed "newKit" import "@celo/contractkit". On the contract i created a withdrawProposal function that deletes submitted proposal.